WIP
Api refactor
